Moved asr1_new_conf to asr1
