--- egs3/librispeech_100/asr1/conf/fbank.conf
@@ -1,2 +1,0 @@
---sample-frequency=16000
---num-mel-bins=80
