--- egs3/librispeech_100/asr1/conf/pbs.conf
@@ -1,11 +1,0 @@
-# Default configuration
-command qsub -V -v PATH -S /bin/bash
-option name=* -N $0
-option mem=* -l mem=$0
-option mem=0          # Do not add anything to qsub_opts
-option num_threads=* -l ncpus=$0
-option num_threads=1  # Do not add anything to qsub_opts
-option num_nodes=* -l nodes=$0:ppn=1
-default gpu=0
-option gpu=0
-option gpu=* -l ngpus=$0
