--- egs3/librispeech_100/asr1/conf/pitch.conf
@@ -1,1 +1,0 @@
---sample-frequency=16000
