--- egs3/librispeech_100/asr1/conf/queue.conf
@@ -1,12 +1,0 @@
-# Default configuration
-command qsub -v PATH -cwd -S /bin/bash -j y -l arch=*64*
-option name=* -N $0
-option mem=* -l mem_free=$0,ram_free=$0
-option mem=0          # Do not add anything to qsub_opts
-option num_threads=* -pe smp $0
-option num_threads=1  # Do not add anything to qsub_opts
-option max_jobs_run=* -tc $0
-option num_nodes=* -pe mpi $0  # You must set this PE as allocation_rule=1
-default gpu=0
-option gpu=0
-option gpu=* -l gpu=$0 -q g.q
