--- egs3/librispeech_100/asr1/conf/slurm.conf
@@ -1,14 +1,0 @@
-# Default configuration
-command sbatch --export=PATH
-option name=* --job-name $0
-option time=* --time $0
-option mem=* --mem-per-cpu $0
-option mem=0
-option num_threads=* --cpus-per-task $0
-option num_threads=1 --cpus-per-task 1
-option num_nodes=* --nodes $0
-default gpu=0
-option gpu=0 --account bbjs-delta-cpu -p cpu
-option gpu=* --account bbjs-delta-gpu -p gpuA40x4,gpuA100x4 --gres=gpu:$0 -c $0  # Recommend allocating more CPU than, or equal to the number of GPU
-# note: the --max-jobs-run option is supported as a special case
-# by slurm.pl and you don't have to handle it in the config file.
