+++ egs3/librispeech_100/asr1/evaluate.yaml
@@ -1,0 +1,1 @@
+conf/inference_beam1.yaml
