--- egs3/librispeech_100/asr1/inference.sh
@@ -1,15 +1,0 @@
-#!/bin/bash
-#SBATCH --nodes=1
-#SBATCH --output=logs/train_global_%j.log
-#SBATCH --ntasks-per-node=1
-#SBATCH --cpus-per-task=16
-#SBATCH --mem=32G
-#SBATCH --time=24:00:00
-#SBATCH -p cpu
-#SBATCH --account bbjs-delta-cpu
-
-. path.sh
-
-export N_GPU=1
-
-srun python ./inference.py --config inference.yaml
