+++ egs3/librispeech_100/asr1/train.yaml
@@ -1,0 +1,1 @@
+conf/train_ctc.yaml
