+++ egs3/librispeech_100/asr1/train_gpu2.yaml
@@ -1,0 +1,1 @@
+conf/train_ctc_gpu2.yaml
