--- egs3/librispeech_100/asr1_new_conf/inference.yaml
@@ -1,1 +1,0 @@
-conf/inference_beam1.yaml
