--- egs3/librispeech_100/asr1_new_conf/train.yaml
@@ -1,1 +1,0 @@
-conf/train_debug.yaml
